Telemetry services implementations

+++ src/Splitio/Telemetry/Domain/HTTPErrors.cs
@@ -6,16 +6,16 @@ namespace Splitio.Telemetry.Domain
     public class HTTPErrors
     {
         [JsonProperty("sp")]
-        public IDictionary<long, long> Splits { get; set; }
+        public IDictionary<int, long> Splits { get; set; }
         [JsonProperty("se")]
-        public IDictionary<long, long> Segments { get; set; }
+        public IDictionary<int, long> Segments { get; set; }
         [JsonProperty("im")]
-        public IDictionary<long, long> Impressions { get; set; }
+        public IDictionary<int, long> Impressions { get; set; }
         [JsonProperty("ev")]
-        public IDictionary<long, long> Events { get; set; }
+        public IDictionary<int, long> Events { get; set; }
         [JsonProperty("to")]
-        public IDictionary<long, long> Token { get; set; }
+        public IDictionary<int, long> Token { get; set; }
         [JsonProperty("te")]
-        public IDictionary<long, long> Telemetry { get; set; }
+        public IDictionary<int, long> Telemetry { get; set; }
     }
 }
